Move metalsmith personal plugins in a different file
